Move unnecessary classes to default factory
