+++ packages/core/src/factory/interface.ts
@@ -1,6 +1,7 @@
+import { Fraction } from '@uniswap/sdk-core'
 import { constants, RpcProvider } from 'starknet'
 
-import { Memecoin } from './memecoin'
+import { BaseMemecoin, Memecoin, MemecoinLaunchData } from '../types/memecoin'
 
 export type FactoryConfig = {
   provider: RpcProvider
@@ -11,4 +12,8 @@ export abstract class FactoryInterface {
   public abstract config: FactoryConfig
 
   public abstract getMemecoin(address: string): Promise<Memecoin | undefined>
+  public abstract getBaseMemecoin(address: string): Promise<BaseMemecoin | undefined>
+  public abstract getMemecoinLaunchData(address: string): Promise<MemecoinLaunchData>
+
+  public abstract getStartingMarketCap(memecoin: Memecoin, quoteTokenPriceAtLaunch?: Fraction): Fraction | undefined
 }
